Refactor: refactorizé todo el codigo existente de css creando variables y eliminando propiedades duplicadas en los estilos.
