+++ src/app/_variables.scss
@@ -28,7 +28,8 @@ $padding-2: 2rem;
 $margin-1: 1rem;
 $margin-2: 2rem;
 // ============================ WIDTH / HEIGHT ====================== //
-$width-100vw: 100vh;
+$width-100vh: 100vh;
+$width-100vw: 100vw;
 $width-100-percent: 100%;
 $width-90-percent: 90%;
 $width-85-percent: 85%;
