--- src/app/docs/docs.module.scss
@@ -1,11 +1,0 @@
-@import "../mixins";
-@import "../variables";
-
-
-.button_wrapper {
-    width: 100%;
-    margin: 0 auto;
-    @include flex-h-center;
-    margin: 2em 0em 2em 0em;
-    gap: 1em;
-}
