+++ src/app/page.module.scss
@@ -2,8 +2,8 @@
 @import "./variables";
 
 .main_wrapper {
-	width: 100%;
-	height: 100%;
+	width: $width-100-percent;
+	height: $width-100-percent;
 	background-image: url(../../public/images/bg/planets.jpg);
 	background-size: cover;
 	background-position: center;
