+++ src/app/snippets/[snippets_tech]/[snippetId]/snippet_chosen.module.scss
@@ -2,7 +2,7 @@
 @import "../../../mixins";
 
 .containerFluid {
-	width: 100%;
+	width: $width-100-percent;
 	padding: 2em;
 }
 
@@ -10,7 +10,7 @@
 	padding: 2em;
 	color: $text-color;
 	background-color: $bg-light-color;
-	width: 90%;
+	width: $width-90-percent;
 	margin: 0 auto;
 	border-radius: 10px;
 }
