--- src/components/statefull/buttons/scroll-button.module.scss
@@ -1,6 +1,0 @@
-@import "../../../../src/app/variables";
-@import "../../../../src/app/mixins";
-
-.scroll_btn {
-	@include scrollTopButton;
-}
