+++ src/components/stateless/HomeBanner/home-page-banner.module.scss
@@ -1,26 +1,33 @@
+@import "../../../app/mixins";
+@import "../../../app/variables";
+
+.image_background,
+.space_background {
+	width: $width-100-percent;
+	height: $width-100vh;
+	background-size: cover;
+	background-repeat: no-repeat;
+}
+
+.space_background,
+.responsive_image,
+.banner_title {
+	position: absolute;
+}
+
 .image_background {
 	position: relative;
-	width: 100%;
-	height: 100vh;
 	background-image: url("../../../../public/images/bg/space-bg.jpg");
 	background-position: center;
-	background-repeat: no-repeat;
-	background-size: cover;
 	background-attachment: fixed;
 	margin-top: 4.7%;
 
 	.space_background {
-		position: absolute;
-		width: 100%;
-		height: 100%;
 		background-image: url("../../../../public/images/bg/main-bg.png");
-		background-position: 50% 50%;
-		background-repeat: no-repeat;
-		background-size: cover;
+		background-position: $width-50-percent $width-50-percent;
 	}
 
 	.responsive_image {
-		position: absolute;
 		width: 250px;
 		height: 250px;
 		object-fit: cover;
@@ -29,9 +36,8 @@
 	}
 
 	.banner_title {
-		position: absolute;
 		z-index: 1;
-		color: white;
+		color: $white-color;
 		top: 550px;
 		left: 500px;
 		font-size: clamp(2rem, 5vw, 5rem);
@@ -72,27 +78,17 @@
 
 @media (max-width: 1300px) {
 	.image_background {
-		position: relative;
 		width: 1300px;
 		height: 700px;
-		background-image: url("../../../../public/images/bg/space-bg.jpg");
-		background-position: 50% 50%;
-		background-repeat: no-repeat;
-		background-size: cover;
-		background-attachment: fixed;
+		background-position: $width-50-percent $width-50-percent;
 		margin-top: 0;
 		.space_background {
-			position: absolute;
 			width: 1300px;
 			height: 700px;
-			background-image: url("../../../../public/images/bg/main-bg.png");
-			background-position: 50% 50%;
-			background-repeat: no-repeat;
-			background-size: cover;
+			background-position: $width-50-percent $width-50-percent;
 		}
 
 		.responsive_image {
-			position: absolute;
 			width: 250px;
 			height: 250px;
 			top: 280px;
@@ -103,28 +99,17 @@
 
 @media (max-width: 1200px) {
 	.image_background {
-		position: relative;
 		width: 1200px;
 		height: 570px;
-
-		background-image: url("../../../../public/images/bg/space-bg.jpg");
-		background-position: 50% 50%;
-		background-repeat: no-repeat;
-		background-size: cover;
-		background-attachment: fixed;
+		background-position: $width-50-percent $width-50-percent;
 		margin-top: 4.7%;
 		.space_background {
-			position: absolute;
 			width: 1200px;
 			height: 570px;
-			background-image: url("../../../../public/images/bg/main-bg.png");
-			background-position: 50% 50%;
-			background-repeat: no-repeat;
-			background-size: cover;
+			background-position: $width-50-percent $width-50-percent;
 		}
 
 		.responsive_image {
-			position: absolute;
 			width: 250px;
 			height: 250px;
 			top: 200px;
@@ -134,28 +119,18 @@
 }
 @media (max-width: 1024px) {
 	.image_background {
-		position: relative;
 		width: 1024px;
 		height: 550px;
 		margin-top: 4.7%;
-		background-image: url("../../../../public/images/bg/space-bg.jpg");
-		background-position: 50% 50%;
-		background-repeat: no-repeat;
-		background-size: cover;
-		background-attachment: fixed;
+		background-position: $width-50-percent $width-50-percent;
 
 		.space_background {
-			position: absolute;
 			width: 1024;
 			height: 550px;
-			background-image: url("../../../../public/images/bg/main-bg.png");
-			background-position: 50% 50%;
-			background-repeat: no-repeat;
-			background-size: cover;
+			background-position: $width-50-percent $width-50-percent;
 		}
 
 		.responsive_image {
-			position: absolute;
 			width: 220px;
 			height: 220px;
 			top: 200px;
@@ -165,28 +140,18 @@
 }
 @media (max-width: 900px) {
 	.image_background {
-		position: relative;
 		width: 900px;
 		height: 500px;
 		margin-top: 4.7%;
-		background-image: url("../../../../public/images/bg/space-bg.jpg");
 		background-position: 10% 90%;
-		background-repeat: no-repeat;
-		background-size: cover;
-		background-attachment: fixed;
 
 		.space_background {
-			position: absolute;
 			width: 900px;
 			height: 500px;
-			background-image: url("../../../../public/images/bg/main-bg.png");
 			background-position: 10% 90%;
-			background-repeat: no-repeat;
-			background-size: cover;
 		}
 
 		.responsive_image {
-			position: absolute;
 			width: 200px;
 			height: 200px;
 			top: 170px;
@@ -196,28 +161,17 @@
 }
 @media (max-width: 700px) {
 	.image_background {
-		position: relative;
 		width: 700px;
 		height: 350px;
 		margin-top: 4.7%;
-		background-image: url("../../../../public/images/bg/space-bg.jpg");
 		background-position: 15% 20%;
-		background-repeat: no-repeat;
-		background-size: cover;
-		background-attachment: fixed;
-
 		.space_background {
-			position: absolute;
 			width: 700px;
 			height: 350px;
-			background-image: url("../../../../public/images/bg/main-bg.png");
 			background-position: 15% 20%;
-			background-repeat: no-repeat;
-			background-size: cover;
 		}
 
 		.responsive_image {
-			position: absolute;
 			width: 150px;
 			height: 150px;
 			top: 150px;
