+++ src/components/stateless/container-button/container_button.module.scss
@@ -2,7 +2,7 @@
 @import "../../../app/mixins";
 
 .container_button {
-    width: 100%;
+    width: $width-100-percent;
     margin: 0 auto;
     @include flex-h-center;
     margin: 2em 0em 2em 0em;
