+++ src/components/stateless/our-team/our_team.module.scss
@@ -2,35 +2,40 @@
 @import "../../../app/variables";
 
 .our_team_container {
-    width: 100%;
-   margin: 0 auto;
-   color: $white-color;
-   padding: 1em 2em;
-   gap: 2em;
-  display: flex;
-  justify-content: center;
+	width: $width-100-percent;
+	margin: 0 auto;
+	color: $white-color;
+	padding: 1em 2em;
+	gap: 2em;
+	display: flex;
+	justify-content: center;
   flex-wrap: wrap;
 }
 
 .our_team_card {
-    @include glass;
-    padding: 1em;
-    border-radius: 10px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 2em;
+	@include glass;
+	padding: 1em;
+	border-radius: 10px;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	padding: 2em;
 }
 
 .our_team_card:hover {
 	background: linear-gradient(to right, #016cec, #00b5f7);
 }
 
+.our_team_card_dashed,
+.our_team_img,
+.our_team_icon {
+    border-radius: $width-50-percent;
+}
+
 .our_team_card_dashed {
     position: relative;
     margin-bottom: 50px;
-    border-radius: 50%;
     border: 4px dashed transparent;
     transition: padding .3s;
     padding: 12px;
@@ -41,15 +46,12 @@
  .our_team_img {
 	width: 120px;
     height: 120px;
-    border-radius: 50%;
 }
-
-
 
 .our_team_icon {
 	position: absolute;
     bottom: 0;
-    left: 50%;
+    left: $width-50-percent;
     transform: translate(-50%,50%);
     display: inline-block;
     width: 60px;
@@ -58,7 +60,6 @@
     text-align: center;
     background: linear-gradient(to right, #016cec, #00b5f7);
     color: #fff;
-    border-radius: 50%;
     font-size: 24px;
 }
